deleted unnecessary folder mysite; moved non-user specific models from auth0login to tradingbot
