+++ backend/auth0login/models.py
@@ -1,7 +1,7 @@
 from django.contrib.auth.models import User
 from django.db import models
 
-from backend.tradingbot.models import Company
+from backend.tradingbot.models import Company, Stock
 
 
 class Credential(models.Model):
@@ -84,23 +84,6 @@ class BotInstance(models.Model):
                f' Portfolio: {self.portfolio.name}'
 
 
-class Stock(models.Model):
-    """Stock of a company"""
-    company = None  # To Be Completed
-    current_price = None  # To Be Completed
-    indicators = None  # To Be Completed
-    Historical_prices = None  # To Be Completed
-    Historical_volatility = None  # To Be Completed
-
-    # Metadata
-    class Meta:
-        ordering = ['company']
-
-    # Methods
-    def __str__(self):
-        return None  # To Be Completed
-
-
 class StockInstance(models.Model):
     """An instance of a stock"""
     portfolio = models.ForeignKey(Portfolio, help_text='Associated portfolio', on_delete=models.CASCADE)
@@ -114,63 +97,3 @@ class StockInstance(models.Model):
     # Methods
     def __str__(self):
         return f'Stock: {str(self.stock)} \n Quantity: {self.quantity} \n Portfolio: {self.portfolio.name}'
-
-
-class News(models.Model):
-    """News of a company"""
-    headline = models.TextField()
-    link = models.URLField(max_length=200)  # default = 200
-    date = models.DateField(auto_now=False, auto_now_add=False)
-
-    # Metadata
-    class Meta:
-        ordering = ['date']
-
-    # Methods
-    def __str__(self):
-        return f'Healine: {str(self.headline)} \n Link: {self.link} \n Date: {self.date}'
-
-
-class Tweets(models.Model):
-    """Tweets/Reddits of a compay"""
-    content = models.TextField()
-    date = models.DateField(auto_now=False, auto_now_add=False)
-
-    # Metadata
-    class Meta:
-        ordering = ['date']
-
-    # Methods
-    def __str__(self):
-        return f'Content: {str(self.content)} \n Date: {self.date}'
-
-
-class Company(models.Model):
-    """Company entity"""
-    name = models.TextField()
-    ticker = models.BigAutoField(primary_key=True)
-    news = models.ManyToManyField(News)
-    tweets = models.ManyToManyField(Tweets)
-
-    # Metadata
-    class Meta:
-        ordering = ['ticker']
-
-    # Methods
-    def __str__(self):
-        return f'Name: {str(self.name)} \n Ticker: {self.ticker}'
-
-
-class Price(models.Model):
-    """Price of a stock"""
-    stock = models.ForeignKey(Stock, help_text='Associated stock', on_delete=models.CASCADE)
-    date = models.DateField(auto_now=False, auto_now_add=False)
-    value = models.DecimalField(max_digits=8, decimal_places=2, help_text='quantity')
-
-    # Metadata
-    class Meta:
-        ordering = ['date']
-
-    # Methods
-    def __str__(self):
-        return f'Stock: {str(self.stock.company)} \n Date: {self.date} \n Value: {self.value}'
